@tohtana 2025-10-02
@eternalNight As some of our CI env use very old PyTorch, they fail with loading deepspeed. Can you add a guard to avoid the error?
I think DeepCompile now works with v2.6 or later, and this PR should work with these versions. But I'm okay to limit it to v2.7 or later if necessary.

@loadams @sfc-gh-truwase This doesn't need to be addressed for this PR, but do we have a plan to update the supported PyTorch version? python tests fail because the PyTorch version there was very old. The top page says
> For full feature support we recommend a version of PyTorch that is >= 1.9 and ideally the latest PyTorch stable release.

but it would be okay to limit to 2.0 or later.